Moved generated types; js_only.
