+++ assets/prototype/domain/shared/types.d.ts
